milan/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
channels==4.0.0
daphne==4.0.0
whitenoise==6.6.0
dj-database-url==2.1.0
